+++ lib/wasix/src/runners/mod.rs
@@ -1,5 +1,3 @@
-mod runner;
-
 #[cfg(feature = "webc_runner_rt_emscripten")]
 pub mod emscripten;
 pub mod wasi;
@@ -7,7 +5,8 @@ mod wasi_common;
 #[cfg(feature = "webc_runner_rt_wcgi")]
 pub mod wcgi;
 
-pub use self::{runner::Runner, wasi_common::MappedCommand};
+pub use self::wasi_common::MappedCommand;
+pub use crate::runtime::runner::Runner;
 
 /// A directory that should be mapped from the host filesystem into a WASI
 /// instance (the "guest").
